ksh93: fix BSD builds due to iconv confusion
For the BSD's we do not want to pass CFLAGS since it includes a
-I/usr/local/incude directive.

This breaks ksh's iconv detection due to the weird way in which iconv
seems to be handled on the BSD's - both a libc impl (preferred), and a
possibly external GNU iconv impl installed in /usr/local.

/usr/local/include is added to CFLAGS by the X11/Motif detection logic
- since that is where all of the needed headers are on the BSDs.

One of the patches from Martijn Decker added CFLAGS to the ksh93 build
CCFLAGS which made this problem show up.

So until/unless that is fixed in ksh93, we will avoid sending
anything to the ksh build system except for SUIDEXECDEFINES